refact: move UlidIdentityGenerator class to utils package
