+++ gateway/src/test/kotlin/com/bibbi/utils/UlidIdentityGeneratorTest.kt
@@ -1,4 +1,4 @@
-package com.bibbi.component
+package com.bibbi.utils
 
 import org.hibernate.validator.internal.util.Contracts.assertNotNull
 import org.hibernate.validator.internal.util.Contracts.assertTrue
